Move CAS1 Booking Cancelled Domain Event Logic
This commit moves the Booking Cancelled Domain Event Logic into the Cas1BookingDomainEventService. This commit only makes essential changes to the shifted code to keep it functioning for traceability.
