+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/BookingServiceTest.kt
@@ -22,7 +22,6 @@ import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.EnumSource
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.http.HttpStatus
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.BookingCancelledEnvelope
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.DestinationProvider
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.PersonReference
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Premises
@@ -94,7 +93,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DestinationPr
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ExtensionEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ExtensionRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.LostBedsRepository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.MetaDataName
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.MoveOnCategoryRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.NonArrivalEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.NonArrivalReasonRepository
@@ -113,7 +111,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserQualifica
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.listeners.BookingListener
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.ApprovedPremisesApplicationStatus
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.DomainEvent
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.PersonInfoResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.problem.ForbiddenProblem
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.problem.InternalServerErrorProblem
@@ -141,7 +138,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Withdrawabl
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.WithdrawalContext
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.WithdrawalTriggeredByUser
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.unit.util.addRoleForUnitTest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.isWithinTheLastMinute
 import java.time.Instant
 import java.time.LocalDate
 import java.time.OffsetDateTime
@@ -2499,23 +2495,11 @@ class BookingServiceTest {
         .withCrn(application.crn)
         .produce()
       every { mockCancellationReasonRepository.findByIdOrNull(reasonId) } returns reason
-      every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
+
+      val cancellationSaveArgument = slot<CancellationEntity>()
+      every { mockCancellationRepository.save(capture(cancellationSaveArgument)) } answers { it.invocation.args[0] as CancellationEntity }
+      every { mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
-
-      val offenderDetails = OffenderDetailsSummaryFactory()
-        .withCrn(bookingEntity.crn)
-        .produce()
-
-      every { mockOffenderService.getOffenderByCrn(bookingEntity.crn, user.deliusUsername) } returns AuthorisableActionResult.Success(offenderDetails)
-
-      val staffUserDetails = StaffUserDetailsFactory().produce()
-
-      every { mockCommunityApiClient.getStaffUserDetails(user.deliusUsername) } returns ClientResult.Success(
-        HttpStatus.OK,
-        staffUserDetails,
-      )
-
       every { mockBookingRepository.findAllByApplication(application) } returns emptyList()
       every { mockApplicationService.updateApprovedPremisesApplicationStatus(any(), any()) } returns Unit
 
@@ -2544,39 +2528,14 @@ class BookingServiceTest {
       assertThat(result.value.notes).isEqualTo("notes")
       assertThat(result.value.booking.status).isEqualTo(BookingStatus.cancelled)
 
-      val domainEventArgument = slot<DomainEvent<BookingCancelledEnvelope>>()
-
       verify(exactly = 1) {
-        mockDomainEventService.saveBookingCancelledEvent(
-          capture(domainEventArgument),
+        mockCas1BookingDomainEventService.bookingCancelled(
+          bookingEntity,
+          user,
+          cancellationSaveArgument.captured,
+          reason,
         )
       }
-
-      val domainEvent = domainEventArgument.captured
-
-      assertThat(domainEvent.applicationId).isEqualTo(application.id)
-      assertThat(domainEvent.crn).isEqualTo(application.crn)
-      assertThat(domainEvent.nomsNumber).isEqualTo(offenderDetails.otherIds.nomsNumber)
-      assertThat(domainEvent.schemaVersion).isEqualTo(2)
-
-      val data = domainEvent.data.eventDetails
-      assertThat(data.applicationId).isEqualTo(application.id)
-      assertThat(data.applicationUrl).isEqualTo("http://frontend/applications/${application.id}")
-      assertThat(data.personReference.crn).isEqualTo(offenderDetails.otherIds.crn)
-      assertThat(data.personReference.noms).isEqualTo(offenderDetails.otherIds.nomsNumber)
-      assertThat(data.deliusEventNumber).isEqualTo(application.eventNumber)
-
-      assertThat(data.premises.id).isEqualTo(premises.id)
-      assertThat(data.premises.name).isEqualTo(premises.name)
-      assertThat(data.premises.apCode).isEqualTo(premises.apCode)
-      assertThat(data.premises.legacyApCode).isEqualTo(premises.qCode)
-      assertThat(data.premises.localAuthorityAreaName).isEqualTo(premises.localAuthorityArea!!.name)
-
-      assertThat(data.cancelledAt).isEqualTo(Instant.parse("2022-08-25T00:00:00.00Z"))
-      assertThat(data.cancelledAtDate).isEqualTo(LocalDate.parse("2022-08-25"))
-      assertThat(data.cancellationReason).isEqualTo(reason.name)
-      assertThat(data.cancellationRecordedAt).isWithinTheLastMinute()
-      assertThat(data.bookingId).isEqualTo(bookingEntity.id)
 
       verify(exactly = 1) {
         mockBookingRepository.save(bookingEntity)
@@ -2608,22 +2567,8 @@ class BookingServiceTest {
 
       every { mockCancellationReasonRepository.findByIdOrNull(reasonId) } returns reason
       every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
+      every { mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
-
-      val offenderDetails = OffenderDetailsSummaryFactory()
-        .withCrn(bookingEntity.crn)
-        .produce()
-
-      every { mockOffenderService.getOffenderByCrn(bookingEntity.crn, user.deliusUsername) } returns AuthorisableActionResult.Success(offenderDetails)
-
-      val staffUserDetails = StaffUserDetailsFactory().produce()
-
-      every { mockCommunityApiClient.getStaffUserDetails(user.deliusUsername) } returns ClientResult.Success(
-        HttpStatus.OK,
-        staffUserDetails,
-      )
-
       every { mockBookingRepository.findAllByApplication(application) } returns emptyList()
       every { mockApplicationService.updateApprovedPremisesApplicationStatus(any(), any()) } returns Unit
       every { mockCas1BookingEmailService.bookingWithdrawn(application, bookingEntity, placementApplication, WithdrawalTriggeredByUser(user)) } returns Unit
@@ -2662,21 +2607,11 @@ class BookingServiceTest {
         .produce()
 
       every { mockCancellationReasonRepository.findByIdOrNull(reasonId) } returns reason
-      every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
+
+      val cancellationSaveArgument = slot<CancellationEntity>()
+      every { mockCancellationRepository.save(capture(cancellationSaveArgument)) } answers { it.invocation.args[0] as CancellationEntity }
+      every { mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
-      val offenderDetails = OffenderDetailsSummaryFactory()
-        .withCrn(bookingEntity.crn)
-        .produce()
-
-      every { mockOffenderService.getOffenderByCrn(bookingEntity.crn, user.deliusUsername) } returns AuthorisableActionResult.Success(offenderDetails)
-
-      val staffUserDetails = StaffUserDetailsFactory().produce()
-
-      every { mockCommunityApiClient.getStaffUserDetails(user.deliusUsername) } returns ClientResult.Success(
-        HttpStatus.OK,
-        staffUserDetails,
-      )
 
       val cancelledAt = LocalDate.parse("2022-08-25")
       val notes = "notes"
@@ -2702,34 +2637,14 @@ class BookingServiceTest {
       assertThat(result.value.booking.status).isEqualTo(BookingStatus.cancelled)
 
       verify(exactly = 1) {
-        mockDomainEventService.saveBookingCancelledEvent(
-          match {
-            val data = it.data.eventDetails
-
-            it.applicationId == application.id &&
-              it.crn == application.crn &&
-              it.nomsNumber == offenderDetails.otherIds.nomsNumber &&
-              it.metadata.contains(MetaDataName.CAS1_CANCELLATION_ID) &&
-              it.metadata.get(MetaDataName.CAS1_CANCELLATION_ID) == result.value.id.toString() &&
-              data.applicationId == application.id &&
-              data.applicationUrl == "http://frontend/applications/${application.id}" &&
-              data.personReference == PersonReference(
-              crn = offenderDetails.otherIds.crn,
-              noms = offenderDetails.otherIds.nomsNumber!!,
-            ) &&
-              data.deliusEventNumber == application.eventNumber &&
-              data.premises == Premises(
-              id = premises.id,
-              name = premises.name,
-              apCode = premises.apCode,
-              legacyApCode = premises.qCode,
-              localAuthorityAreaName = premises.localAuthorityArea!!.name,
-            ) &&
-              data.cancelledAt == Instant.parse("2022-08-25T00:00:00.00Z") &&
-              data.cancellationReason == reason.name
-          },
+        mockCas1BookingDomainEventService.bookingCancelled(
+          bookingEntity,
+          user,
+          cancellationSaveArgument.captured,
+          reason,
         )
       }
+
       verify(exactly = 1) {
         mockBookingRepository.save(bookingEntity)
       }
@@ -2749,7 +2664,6 @@ class BookingServiceTest {
 
       every { mockCancellationReasonRepository.findByIdOrNull(reasonId) } returns reason
       every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
       val offenderDetails = OffenderDetailsSummaryFactory()
         .withCrn(bookingEntity.crn)
@@ -2788,7 +2702,7 @@ class BookingServiceTest {
       assertThat(result.value.booking.status).isEqualTo(BookingStatus.cancelled)
 
       verify(exactly = 0) {
-        mockDomainEventService.saveBookingCancelledEvent(any())
+        mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any())
       }
       verify(exactly = 1) {
         mockBookingRepository.save(bookingEntity)
@@ -2943,21 +2857,8 @@ class BookingServiceTest {
 
       every { mockCancellationReasonRepository.findByIdOrNull(reasonId) } returns reason
       every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
+      every { mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
-
-      val offenderDetails = OffenderDetailsSummaryFactory()
-        .withCrn(bookingEntity.crn)
-        .produce()
-
-      every { mockOffenderService.getOffenderByCrn(bookingEntity.crn, user.deliusUsername) } returns AuthorisableActionResult.Success(offenderDetails)
-
-      val staffUserDetails = StaffUserDetailsFactory().produce()
-
-      every { mockCommunityApiClient.getStaffUserDetails(user.deliusUsername) } returns ClientResult.Success(
-        HttpStatus.OK,
-        staffUserDetails,
-      )
 
       every {
         mockApplicationService.updateApprovedPremisesApplicationStatus(
@@ -3038,21 +2939,8 @@ class BookingServiceTest {
 
       every { mockCancellationReasonRepository.findByIdOrNull(reasonId) } returns reason
       every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
+      every { mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
-
-      val offenderDetails = OffenderDetailsSummaryFactory()
-        .withCrn(bookingEntity.crn)
-        .produce()
-
-      every { mockOffenderService.getOffenderByCrn(bookingEntity.crn, user.deliusUsername) } returns AuthorisableActionResult.Success(offenderDetails)
-
-      val staffUserDetails = StaffUserDetailsFactory().produce()
-
-      every { mockCommunityApiClient.getStaffUserDetails(user.deliusUsername) } returns ClientResult.Success(
-        HttpStatus.OK,
-        staffUserDetails,
-      )
 
       every {
         mockApplicationService.updateApprovedPremisesApplicationStatus(
@@ -3116,7 +3004,7 @@ class BookingServiceTest {
 
       every { mockCancellationReasonRepository.findByIdOrNull(CancellationReasonRepository.CAS1_RELATED_PLACEMENT_APP_WITHDRAWN_ID) } returns reason
       every { mockCancellationRepository.save(any()) } answers { it.invocation.args[0] as CancellationEntity }
-      every { mockDomainEventService.saveBookingCancelledEvent(any()) } just Runs
+      every { mockCas1BookingDomainEventService.bookingCancelled(any(), any(), any(), any()) } just Runs
       every { mockBookingRepository.save(any()) } answers { it.invocation.args[0] as BookingEntity }
 
       val offenderDetails = OffenderDetailsSummaryFactory()
